pyproject.toml
[project]
name = "entity-plugin-template"
version = "0.1.0"
description = "Template for creating Entity Framework plugins"
authors = [
    {name = "Entity Framework Team", email = "team@entity-framework.org"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "entity-core>=0.0.1",
    "pydantic>=2.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I"]
